backend/requirements.txt

fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pinecone-client==2.2.4
openai==1.3.5
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
alembic==1.12.0
